A HTTP transport was added.
Also, now both Websocket and HTTP transports are export MAKE-CLACK-APP
functions. This is useful, when you want to combine api application
with some other Clack applications using :mount middleware.

HTTP Transport is useful, when you don't want to keep persistent websocket or TCP
connections for each client, because it can be used as a DoS attack on the server.